refactor: keep the shared spec limit in a module that imports nothing
Putting the constant in the orchestrate package's _common charged both Studio
services for it. Not because _common is heavy — importing any submodule of
lionagi.cli.orchestrate runs that package's __init__ first, and that is the
expensive part. Measured on fresh processes, three runs each, cumulative import
time for the service module:

  playbooks   365735 us -> 219766 us
  schedules   426685 us -> 345597 us

The left column is main, the right is this branch reading the constant from a
top-level leaf that imports nothing. Both are now at or below where they started.

No behaviour changes; the value and every check that reads it are unchanged.
